requirements.txt
Django
Jinja2==2.5.5
boto
-e git+https://github.com/sunlightlabs/django-mediasync.git/#egg=mediasync
Djinja

#http://downloads.sourceforge.net/project/plastex/plastex/rel-0-9-2/plastex-0.9.2.tgz?r=http%3A%2F%2Fsourceforge.net%2Fprojects%2Fplastex%2Ffiles%2Fplastex%2Frel-0-9-2%2F&ts=1304424112&use_mirror=sunet

django-piston
django-social-auth
django-picklefield

PIL